refactor: move gRPC service generators
